Optimization to choose a flight with live ads
This fixes a known issue with the flight decision backend.
It is possible to have the backend return no ad at all.
This happens when there's a live flight with live ads
but the flight has only non-live ads of the chosen type.

This can happen for example if an advertiser was testing
text-only and text+image ads but then decided to disable
all ads of one type.