Don’t index `ad_type` on the AdBase
This caused a big downtime on migration trying to write to our huge View table.
We never want to query on the ad_type, at least not yet.
We can do a concurrent index at some point in the future if we need to,
but we don’t want it on the migration.