Make Offers nullable to track fill rate
This allows us to start storing data about requests we don’t make Offers for.
We need to start adding this data into our aggregated reports,
but for now we can at least start storing it.

I think this migration should be pretty safe,
but we probably want to be ready to cancel it if not.