Add uplift reporting
This reuses the All Publishers report,
and just changes the index on the total.
This actually worked pretty cleanly.

Another option for this would be to add `uplift` to the AdImpression index.
The data is all the same,
and it's just an additional boolean that we'd pivot on.
We can always go that way if we want,
but this approach seemed simplest for now.