Remove suggested_capabilities from the Stripe API
I'm attempting to get it to properly set the `recipient` account type,
which will allow us to send money to folks overseas.
Currently all our accounts are set to `full`,
which disables this ability.

I've changed the settings in the dashboard,
but I believe they are not being respected becuase of this line in the docs:

> You can override the Configuration settings for an individual account by specifying its capabilities and service agreement type with the Accounts API.

https://stripe.com/docs/connect/service-agreement-types#choosing-type