Pace ads by a custom interval
Currently, we pace ads a day at a time.
So if a campaign is 1000 clicks over 10 days,
we will aim to space it out so the campaign gets 100 clicks per day.
Switching to a smaller interval (say, an hour) will have even more
benefits like pacing the campaign more evenly across regions
(not missing any because that region is asleep when the day resets).