Add more precision to the optimized daily tables
This is a debatable change and I think it's worth considering
rejecting this change. The daily rollup tables exposed
a slight difference in how we calculate advertiser daily spend
or publisher daily earn. Due to precision, these are rounded
to the nearest $0.01 on a daily basis and that value is saved
in the database for reporting.

From a spend/earn perspective, I don't think it will really
matter but it could lead to some off by a cent type of "issues".

To illustrate this issue, consider an $0.10 campaign over 11 days.
We will always fulfill a certain number of impressions over the life
of the campaign and we will aim for $0.1/11 per day.
The database will store $0.01 per day when in reality we actually
fulfilled $0.091 per day. If we then add up the days, it will appear
as if the campaign was $0.11 instead of $0.10.

With this change, we will round to the nearest 4 decimals
before storing in the database. Reporting will still round to 2 decimals
but when adding up the days, it will use the full 4 decimals.