Ignore publisher mobile traffic option
We can disable mobile traffic from being displayed on a publisher.
This is useful when a publisher has a large mobile segment
of their audience that isn't a fit for our network.

If a publisher's mobile audience is ignored,
there will be a message on the publisher settings page
with this information.