Fix view/click date to date to offer
There's a small date boundary bug where an can be requested on one day
but viewed/clicked on the following day. This causes some of our
aggregations to be off by 1. This will fix this such that the date where
the ad is requested is the date where the view/click is billed.