Bugfix last analyzed date on analyzer
We have some URLs in our DB that haven't been analyzed. These can be
created by `daily_visited_urls_aggregation`. This causes the
`analyze_url` task to fail when it is called on one of those which it
currently never is automatically.

I'm going to have a separate PR to start analyzing those.